requirements.txt
absl-py==0.9.0
alembic==1.4.1
appdirs==1.4.3
asgiref==3.2.7
astor==0.8.1
atpublic==1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
colorama==0.4.3
configobj==5.0.6
configparser==5.0.0
cycler==0.10.0
databricks-cli==0.10.0
decorator==4.4.2
deeplift==0.6.10.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.5.0
Django==3.0.5
djangorestframework==3.11.0
docker==4.2.0
dpath==2.0.1
dvc==0.92.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
flatten-json==0.1.7
flufl.lock==3.2
funcy==1.14
future==0.18.2
gast==0.3.3
gitdb==4.0.2
GitPython==3.1.0
google-pasta==0.2.0
gorilla==0.3.0
grandalf==0.6
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
hdfviewer==0.11.0
humanize==2.2.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
importlib-resources==1.4.0
imutils==0.5.3
inflect==3.0.2
innvestigate==1.0.8
ipykernel==5.2.1
ipympl==0.5.6
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.4
jsonpath-ng==1.5.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.0
jupyterlab-server==1.1.1
kaggle==1.5.6
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mkl-service==2.3.0
mlflow==1.7.2
more-itertools==8.2.0
nanotime==0.5.2
nbconvert==5.6.1
nbformat==5.0.5
networkx==2.3
notebook==6.0.3
numpy==1.18.2
olefile==0.46
opencv-contrib-python==4.2.0.32
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.7.0
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
ply==3.11
progressbar2==3.50.1
prometheus-client==0.7.1
prometheus-flask-exporter==0.13.0
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pycparser==2.20
pydot==1.4.1
Pygments==2.6.1
pygtrie==2.3.2
pyparsing==2.4.6
pyrsistent==0.16.0
pytest==5.4.1
python-dateutil==2.8.0
python-editor==1.0.4
python-slugify==4.0.0
python-utils==2.4.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
querystring-parser==1.2.4
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
shortuuid==1.0.1
simplejson==3.17.0
six==1.14.0
sklearn==0.0
smmap==3.0.1
SQLAlchemy==1.3.13
sqlparse==0.3.1
stevedore==1.32.0
tabulate==0.8.7
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
texttable==1.6.2
torch==1.1.0
torchvision==0.3.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
treelib==1.6.1
urllib3==1.24.3
virtualenv==20.0.15
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
vis==0.0.5
voluptuous==0.11.7
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zc.lockfile==2.0
zipp==3.1.0
